backend/requirements.txt
django
djangorestframework
pyyaml
requests
django-cors-headers
pymongo[srv]==3.12.3